+++ pkg/injector/patcher/patcher.go
@@ -20,7 +20,7 @@ const (
 
 // PatchOperation represents a discreet change to be applied to a Kubernetes resource.
 type PatchOperation struct {
-	Op    string      `json:"op"`
-	Path  string      `json:"path"`
-	Value interface{} `json:"value,omitempty"`
+	Op    string `json:"op"`
+	Path  string `json:"path"`
+	Value any    `json:"value,omitempty"`
 }
