+++ pkg/injector/service/injector_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 The Dapr Authors
+Copyright 2023 The Dapr Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -14,26 +14,19 @@ limitations under the License.
 package service
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
-	"io"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	v1 "k8s.io/api/admission/v1"
-	authenticationv1 "k8s.io/api/authentication/v1"
+	admissionv1 "k8s.io/api/admission/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/util/uuid"
 	kubernetesfake "k8s.io/client-go/kubernetes/fake"
 
-	"github.com/dapr/dapr/pkg/client/clientset/versioned/fake"
 	"github.com/dapr/dapr/pkg/injector/annotations"
 	"github.com/dapr/dapr/pkg/injector/namespacednamematcher"
 	"github.com/dapr/dapr/pkg/injector/sidecar"
@@ -376,7 +369,7 @@ func TestGetAppIDFromRequest(t *testing.T) {
 	})
 
 	t.Run("can handle empty admissionrequest object", func(t *testing.T) {
-		fakeReq := &v1.AdmissionRequest{}
+		fakeReq := &admissionv1.AdmissionRequest{}
 		appID := getAppIDFromRequest(fakeReq)
 		assert.Equal(t, "", appID)
 	})
@@ -390,7 +383,7 @@ func TestGetAppIDFromRequest(t *testing.T) {
 			},
 		}
 		rawBytes, _ := json.Marshal(fakePod)
-		fakeReq := &v1.AdmissionRequest{
+		fakeReq := &admissionv1.AdmissionRequest{
 			Object: runtime.RawExtension{
 				Raw: rawBytes,
 			},
@@ -398,264 +391,6 @@ func TestGetAppIDFromRequest(t *testing.T) {
 		appID := getAppIDFromRequest(fakeReq)
 		assert.Equal(t, "fakeID", appID)
 	})
-}
-
-func TestHandleRequest(t *testing.T) {
-	authID := "test-auth-id"
-
-	i, err := NewInjector([]string{authID}, Config{
-		TLSCertFile:                       "test-cert",
-		TLSKeyFile:                        "test-key",
-		SidecarImage:                      "test-image",
-		Namespace:                         "test-ns",
-		AllowedServiceAccountsPrefixNames: "vc-proj*:sa-dev*,vc-all-allowed*:*",
-	}, fake.NewSimpleClientset(), kubernetesfake.NewSimpleClientset())
-
-	assert.NoError(t, err)
-	injector := i.(*injector)
-
-	podBytes, _ := json.Marshal(corev1.Pod{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "Pod",
-			APIVersion: "v1",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:              "test-app",
-			Namespace:         "default",
-			CreationTimestamp: metav1.Time{Time: time.Now()},
-			Labels: map[string]string{
-				"app": "test-app",
-			},
-			Annotations: map[string]string{
-				"dapr.io/enabled":  "true",
-				"dapr.io/app-id":   "test-app",
-				"dapr.io/app-port": "3000",
-			},
-		},
-		Spec: corev1.PodSpec{
-			Containers: []corev1.Container{
-				{
-					Name:  "main",
-					Image: "docker.io/app:latest",
-				},
-			},
-		},
-	})
-
-	testCases := []struct {
-		testName         string
-		request          v1.AdmissionReview
-		contentType      string
-		expectStatusCode int
-		expectPatched    bool
-	}{
-		{
-			"TestSidecarInjectSuccess",
-			v1.AdmissionReview{
-				Request: &v1.AdmissionRequest{
-					UID:       uuid.NewUUID(),
-					Kind:      metav1.GroupVersionKind{Group: "", Version: "v1", Kind: "Pod"},
-					Name:      "test-app",
-					Namespace: "test-ns",
-					Operation: "CREATE",
-					UserInfo: authenticationv1.UserInfo{
-						Groups: []string{systemGroup},
-					},
-					Object: runtime.RawExtension{Raw: podBytes},
-				},
-			},
-			runtime.ContentTypeJSON,
-			http.StatusOK,
-			true,
-		},
-		{
-			"TestSidecarInjectWrongContentType",
-			v1.AdmissionReview{},
-			runtime.ContentTypeYAML,
-			http.StatusUnsupportedMediaType,
-			true,
-		},
-		{
-			"TestSidecarInjectInvalidKind",
-			v1.AdmissionReview{
-				Request: &v1.AdmissionRequest{
-					UID:       uuid.NewUUID(),
-					Kind:      metav1.GroupVersionKind{Group: "", Version: "v1", Kind: "Deployment"},
-					Name:      "test-app",
-					Namespace: "test-ns",
-					Operation: "CREATE",
-					UserInfo: authenticationv1.UserInfo{
-						Groups: []string{systemGroup},
-					},
-					Object: runtime.RawExtension{Raw: podBytes},
-				},
-			},
-			runtime.ContentTypeJSON,
-			http.StatusOK,
-			false,
-		},
-		{
-			"TestSidecarInjectGroupsNotContains",
-			v1.AdmissionReview{
-				Request: &v1.AdmissionRequest{
-					UID:       uuid.NewUUID(),
-					Kind:      metav1.GroupVersionKind{Group: "", Version: "v1", Kind: "Pod"},
-					Name:      "test-app",
-					Namespace: "test-ns",
-					Operation: "CREATE",
-					UserInfo: authenticationv1.UserInfo{
-						Groups: []string{"system:kubelet"},
-					},
-					Object: runtime.RawExtension{Raw: podBytes},
-				},
-			},
-			runtime.ContentTypeJSON,
-			http.StatusOK,
-			false,
-		},
-		{
-			"TestSidecarInjectUIDContains",
-			v1.AdmissionReview{
-				Request: &v1.AdmissionRequest{
-					UID:       uuid.NewUUID(),
-					Kind:      metav1.GroupVersionKind{Group: "", Version: "v1", Kind: "Pod"},
-					Name:      "test-app",
-					Namespace: "test-ns",
-					Operation: "CREATE",
-					UserInfo: authenticationv1.UserInfo{
-						UID: authID,
-					},
-					Object: runtime.RawExtension{Raw: podBytes},
-				},
-			},
-			runtime.ContentTypeJSON,
-			http.StatusOK,
-			true,
-		},
-		{
-			"TestSidecarInjectUIDNotContains",
-			v1.AdmissionReview{
-				Request: &v1.AdmissionRequest{
-					UID:       uuid.NewUUID(),
-					Kind:      metav1.GroupVersionKind{Group: "", Version: "v1", Kind: "Pod"},
-					Name:      "test-app",
-					Namespace: "test-ns",
-					Operation: "CREATE",
-					UserInfo: authenticationv1.UserInfo{
-						UID: "auth-id-123",
-					},
-					Object: runtime.RawExtension{Raw: podBytes},
-				},
-			},
-			runtime.ContentTypeJSON,
-			http.StatusOK,
-			false,
-		},
-		{
-			"TestSidecarInjectEmptyPod",
-			v1.AdmissionReview{
-				Request: &v1.AdmissionRequest{
-					UID:       uuid.NewUUID(),
-					Kind:      metav1.GroupVersionKind{Group: "", Version: "v1", Kind: "Pod"},
-					Name:      "test-app",
-					Namespace: "test-ns",
-					Operation: "CREATE",
-					UserInfo: authenticationv1.UserInfo{
-						Groups: []string{systemGroup},
-					},
-					Object: runtime.RawExtension{Raw: nil},
-				},
-			},
-			runtime.ContentTypeJSON,
-			http.StatusOK,
-			false,
-		},
-		{
-			"TestSidecarInjectUserInfoMatchesServiceAccountPrefix",
-			v1.AdmissionReview{
-				Request: &v1.AdmissionRequest{
-					UID:       uuid.NewUUID(),
-					Kind:      metav1.GroupVersionKind{Group: "", Version: "v1", Kind: "Pod"},
-					Name:      "test-app",
-					Namespace: "test-ns",
-					Operation: "CREATE",
-					UserInfo: authenticationv1.UserInfo{
-						Username: "system:serviceaccount:vc-project-star:sa-dev-team-usa",
-					},
-					Object: runtime.RawExtension{Raw: podBytes},
-				},
-			},
-			runtime.ContentTypeJSON,
-			http.StatusOK,
-			true,
-		},
-		{
-			"TestSidecarInjectUserInfoMatchesAllAllowedServiceAccountPrefix",
-			v1.AdmissionReview{
-				Request: &v1.AdmissionRequest{
-					UID:       uuid.NewUUID(),
-					Kind:      metav1.GroupVersionKind{Group: "", Version: "v1", Kind: "Pod"},
-					Name:      "test-app",
-					Namespace: "test-ns",
-					Operation: "CREATE",
-					UserInfo: authenticationv1.UserInfo{
-						Username: "system:serviceaccount:vc-all-allowed-project:dapr",
-					},
-					Object: runtime.RawExtension{Raw: podBytes},
-				},
-			},
-			runtime.ContentTypeJSON,
-			http.StatusOK,
-			true,
-		},
-		{
-			"TestSidecarInjectUserInfoNotMatchesServiceAccountPrefix",
-			v1.AdmissionReview{
-				Request: &v1.AdmissionRequest{
-					UID:       uuid.NewUUID(),
-					Kind:      metav1.GroupVersionKind{Group: "", Version: "v1", Kind: "Pod"},
-					Name:      "test-app",
-					Namespace: "test-ns",
-					Operation: "CREATE",
-					UserInfo: authenticationv1.UserInfo{
-						Username: "system:serviceaccount:vc-bad-project-star:sa-dev-team-usa",
-					},
-					Object: runtime.RawExtension{Raw: podBytes},
-				},
-			},
-			runtime.ContentTypeJSON,
-			http.StatusOK,
-			false,
-		},
-	}
-
-	ts := httptest.NewServer(http.HandlerFunc(injector.handleRequest))
-	defer ts.Close()
-
-	for _, tc := range testCases {
-		tc := tc
-		t.Run(tc.testName, func(t *testing.T) {
-			requestBytes, err := json.Marshal(tc.request)
-			assert.NoError(t, err)
-
-			resp, err := http.Post(ts.URL, tc.contentType, bytes.NewBuffer(requestBytes))
-			assert.NoError(t, err)
-			defer resp.Body.Close()
-
-			assert.Equal(t, tc.expectStatusCode, resp.StatusCode)
-
-			if resp.StatusCode == http.StatusOK {
-				body, err := io.ReadAll(resp.Body)
-				assert.NoError(t, err)
-
-				var ar v1.AdmissionReview
-				err = json.Unmarshal(body, &ar)
-				assert.NoError(t, err)
-
-				assert.Equal(t, tc.expectPatched, len(ar.Response.Patch) > 0)
-			}
-		})
-	}
 }
 
 func TestAllowedControllersServiceAccountUID(t *testing.T) {
